feat: en backend service implementation and add more explaination to the Readme.md
