+++ transfer/transfer-06-http-data-flow/backend-service/build.gradle.kts
@@ -19,15 +19,6 @@ plugins {
 val edcGroupId = "org.eclipse.edc"
 val edcVersion = "0.0.1-milestone-7"
 
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation("org.junit.jupiter:junit-jupiter-api:5.9.2")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.9.2")
-}
-
 tasks.withType<Jar> {
     manifest {
         attributes["Main-Class"] = "org.eclipse.edc.EdcHttpServer"
