+++ transfer/transfer-06-http-data-flow/backend-service/src/main/java/org/eclipse/edc/BackendService.java
@@ -15,9 +15,10 @@
 package org.eclipse.edc;
 
 import com.sun.net.httpserver.HttpServer;
-import org.eclipse.edc.handler.HttpServerFactory;
 import org.eclipse.edc.handler.ReceiverHandler;
 
+import java.io.IOException;
+import java.net.InetSocketAddress;
 import java.util.Optional;
 
 public class BackendService {
@@ -26,10 +27,18 @@ public class BackendService {
 
     public static void main(String[] args) {
         int port = Integer.parseInt(Optional.ofNullable(System.getenv(HTTP_PORT)).orElse("4000"));
-        HttpServer server = HttpServerFactory.createHttpServer(port);
+        var server = createHttpServer(port);
         server.createContext("/receiver", new ReceiverHandler());
         server.setExecutor(null);
         server.start();
         System.out.println("server started at " + port);
+    }
+    
+    private static HttpServer createHttpServer(int port) {
+        try {
+            return HttpServer.create(new InetSocketAddress(port), 0);
+        } catch (IOException e) {
+            throw new RuntimeException("Unable to start server at port " + port, e);
+        }
     }
 }
