--- transfer/transfer-06-http-data-flow/backend-service/src/main/java/org/eclipse/edc/exception/UnableToStartServerException.java
@@ -1,26 +1,0 @@
-/*
- *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.exception;
-
-public class UnableToStartServerException extends RuntimeException {
-
-    public UnableToStartServerException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UnableToStartServerException(String message) {
-        super(message);
-    }
-}
