+++ transfer/transfer-06-http-data-flow/backend-service/src/main/java/org/eclipse/edc/handler/ReceiverHandler.java
@@ -16,38 +16,39 @@ package org.eclipse.edc.handler;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import org.eclipse.edc.parser.DefaultHttpQueryParser;
 import org.eclipse.edc.parser.HttpQueryParserBase;
 
 import java.io.IOException;
-import java.io.OutputStream;
 import java.net.URI;
 import java.util.HashMap;
-import java.util.Map;
 
+/**
+ * A handler which is invoked to process HTTP exchanges.
+ * Each HTTP exchange is handled by one of these handlers.
+ */
 public class ReceiverHandler implements HttpHandler {
 
-    private final HttpQueryParserBase queryParser = new DefaultHttpQueryParser();
+    private final HttpQueryParserBase queryParser = new HttpQueryParserBase();
 
+    /**
+     * Handle the given request and generate an appropriate response.
+     * See {@link HttpExchange} for a description of the steps
+     * involved in handling an exchange.
+     *
+     * @param exchange the exchange containing the request from the
+     *                 client and used to send the response
+     * @throws NullPointerException if exchange is {@code null}
+     * @throws IOException          if an I/O error occurs
+     */
     @Override
     public void handle(HttpExchange exchange) throws IOException {
         // parse request
-        Map<String, Object> parameters = new HashMap<>();
+        var parameters = new HashMap<>();
         URI requestedUri = exchange.getRequestURI();
         String query = requestedUri.getRawQuery();
         queryParser.parseQuery(query, parameters);
 
         System.out.println("Request Body: " + new String(exchange.getRequestBody().readAllBytes()));
-
-        // send response
-        StringBuilder response = new StringBuilder();
-        for (var entry : parameters.entrySet()) {
-            response.append(entry.getKey()).append(" = ").append(parameters.get(entry.getKey())).append("\n");
-        }
-        exchange.sendResponseHeaders(200, response.length());
-        OutputStream os = exchange.getResponseBody();
-        os.write(response.toString().getBytes());
-
-        os.close();
+        exchange.sendResponseHeaders(200, 0);
     }
 }
