+++ transfer/transfer-06-http-data-flow/backend-service/src/main/java/org/eclipse/edc/parser/HttpQueryParserBase.java
@@ -15,8 +15,47 @@
 package org.eclipse.edc.parser;
 
 import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
-public interface HttpQueryParserBase {
-    void parseQuery(String query, Map<String, Object> parameters) throws UnsupportedEncodingException;
+public class HttpQueryParserBase {
+    public void parseQuery(String query, Map<Object, Object> parameters) throws UnsupportedEncodingException {
+        if (query == null) {
+            return;
+        }
+
+        String[] pairs = query.split("&");
+        for (String pair : pairs) {
+            String[] param = pair.split("=");
+            String key = null;
+            String value = null;
+            if (param.length > 0) {
+                key = URLDecoder.decode(param[0],
+                        System.getProperty("file.encoding"));
+            }
+
+            if (param.length > 1) {
+                value = URLDecoder.decode(param[1],
+                        System.getProperty("file.encoding"));
+            }
+
+            if (parameters.containsKey(key)) {
+                var obj = parameters.get(key);
+                if (obj instanceof List<?>) {
+                    List<String> values = (List<String>) obj;
+                    values.add(value);
+
+                } else if (obj instanceof String) {
+                    List<String> values = new ArrayList<>();
+                    values.add((String) obj);
+                    values.add(value);
+                    parameters.put(key, values);
+                }
+            } else {
+                parameters.put(key, value);
+            }
+        }
+    }
 }
